criando cache local em music-service e melhorando a segurança dos mét…

+++ music-service/src/main/java/com/ms/musicservice/service/MusicService.java
@@ -8,9 +8,10 @@ import com.ms.dtos.music.MusicPayloadDTO;
 import com.ms.musicservice.dtos.CreateMusicDTO;
 import com.ms.musicservice.dtos.MusicResponseDTO;
 import com.ms.musicservice.dtos.UpdateMusicDTO;
+import com.ms.musicservice.exceptions.ForbiddenException;
 import com.ms.musicservice.exceptions.MusicAlreadyExistsException;
 import com.ms.musicservice.exceptions.MusicNotFoundException;
-import com.ms.musicservice.messaging.MusicProducer;
+import com.ms.musicservice.messaging.producer.MusicProducer;
 import com.ms.musicservice.model.MusicEntity;
 import com.ms.musicservice.repository.MusicRepository;
 import lombok.RequiredArgsConstructor;
@@ -37,13 +38,15 @@ public class MusicService {
     private final MusicRepository musicRepository;
     private final ModelMapper modelMapper;
     private final MusicProducer musicProducer;
+    private final ArtistCacheService artistCacheService;
     private static final Logger logger = LoggerFactory.getLogger(MusicService.class);
     public static final String CACHE_MUSIC="music";
     private static final String CACHE_MUSIC_LIST= "musicList";
 
     @Transactional
     @CacheEvict(value = CACHE_MUSIC_LIST, allEntries = true)
-    public MusicResponseDTO createMusic(UUID artistId, CreateMusicDTO createMusicDTO){
+    public MusicResponseDTO createMusic(UUID userId, CreateMusicDTO createMusicDTO){
+        UUID artistId = artistCacheService.getArtistId(userId);
         logger.info("creating a music. ArtistId: {} | Title: {}", artistId, createMusicDTO.title());
         if(musicRepository.existsByTitleAndArtistId(createMusicDTO.title(), artistId)){
             throw new MusicAlreadyExistsException("A music with the title: " + createMusicDTO.title() + " already exists for this artist");
@@ -55,7 +58,7 @@ public class MusicService {
                 savedMusic.getGenre().toString(), savedMusic.getAlbum());
 
         //send message
-        createMusicEventDto(EventAction.CREATED, musicPayloadDTO );
+        publishMusicEvent(EventAction.CREATED, musicPayloadDTO );
 
         return modelMapper.map(savedMusic, MusicResponseDTO.class);
     }
@@ -66,7 +69,7 @@ public class MusicService {
         var music = findMusicOrThrow(musicId);
         return modelMapper.map(music, MusicResponseDTO.class);
     }
-    @Cacheable(value = CACHE_MUSIC_LIST, key = "'page: ' + #pageable.pageNumber + ':size:' + #pageable.pageSize")
+    @Cacheable(value = CACHE_MUSIC_LIST, key = "#pageable.pageNumber + ':' + #pageable.pageSize + ':' + #pageable.sort")
     @Transactional(readOnly = true)
     public Page<MusicResponseDTO> getAllMusics(Pageable pageable){
         logger.info("get all musics");
@@ -77,31 +80,38 @@ public class MusicService {
             @CacheEvict(value = CACHE_MUSIC, key = "#musicId"),
             @CacheEvict(value = CACHE_MUSIC_LIST, allEntries = true)
     })
-    public void deleteMusic(UUID musicId){
+    public void deleteMusic( UUID userId, UUID musicId){
         logger.info("deleting an music with id: {}", musicId);
         var music = findMusicOrThrow(musicId);
-
+        UUID artistId = artistCacheService.getArtistId(userId);
+        validatePermission(artistId, music);
         musicRepository.delete(music);
 
         //send message
         MusicPayloadDTO musicPayloadDTO = MusicPayloadDTO.forDelete(music.getId(), music.getArtistId());
-        createMusicEventDto(EventAction.DELETED, musicPayloadDTO);
+        publishMusicEvent(EventAction.DELETED, musicPayloadDTO);
     }
     @Transactional
     @Caching(evict = {
             @CacheEvict(value = CACHE_MUSIC, key = "#musicId"),
             @CacheEvict(value = CACHE_MUSIC_LIST, allEntries = true)
     })
-    public MusicResponseDTO updateMusic(UUID musicId, UpdateMusicDTO updateMusicDTO){
+    public MusicResponseDTO updateMusic(UUID userId, UUID musicId, UpdateMusicDTO updateMusicDTO){
         logger.info("updating a music with id: {}", musicId);
         var music = findMusicOrThrow(musicId);
+        UUID artistId = artistCacheService.getArtistId(userId);
+        validatePermission(artistId,music);
+        if (!music.getTitle().equals(updateMusicDTO.title())
+                && musicRepository.existsByTitleAndArtistId(updateMusicDTO.title(), artistId)) {
+            throw new MusicAlreadyExistsException("A music with the title '" + updateMusicDTO.title() + "' already exists for this artist.");
+        }
         modelMapper.map(updateMusicDTO, music);
         var savedMusic = musicRepository.save(music);
 
         //send message
         MusicPayloadDTO musicPayloadDTO = new MusicPayloadDTO(savedMusic.getId(), savedMusic.getTitle(), savedMusic.getLyrics(),savedMusic.getArtistId(),
                 savedMusic.getGenre().toString(), savedMusic.getAlbum());
-        createMusicEventDto(EventAction.UPDATED, musicPayloadDTO );
+        publishMusicEvent(EventAction.UPDATED, musicPayloadDTO );
 
 
         return modelMapper.map(savedMusic, MusicResponseDTO.class);
@@ -110,8 +120,13 @@ public class MusicService {
         return musicRepository.findById(musicId).orElseThrow(() -> new MusicNotFoundException("Music with id: " + musicId + " not found"));
     }
 
-    private void createMusicEventDto(EventAction eventAction, MusicPayloadDTO musicPayloadDTO){
+    private void publishMusicEvent(EventAction eventAction, MusicPayloadDTO musicPayloadDTO){
         MusicEventDTO eventDTO = new MusicEventDTO(eventAction, musicPayloadDTO, LocalDateTime.now());
         musicProducer.handleMusicEvent(eventDTO);
+    }
+    private void validatePermission(UUID artistId, MusicEntity musicEntity){
+        if(!musicEntity.getArtistId().equals(artistId)){
+            throw new ForbiddenException("You do not have permission to access this resource.");
+        }
     }
 }
